Move health checks to a separate function
Now, regarding to the template file, we have to write checks twice: 
- In `@test "install from directory"`
- And in `@test "install from release"` - the same lines.

Will be good to move that to a separate function, to not have two places to manage them. 

So, I've created a function `health_checks()` that contains only a single instance of checks.

And have moved it to the top, because it will be edited more frequiently, than test functions.