Update getting started section point 6 in the README.md
At first i wanted to clarify the details about how to test an add-on locally. I've tried that for the first time today. i've read the documentation on ddev.readthedocs as well as the readme for this repo but each of them wasn't that clear. @rfay helped to clarify. so i thought it would make sense to clarify the process within the addon-template readme as well. aside that i've suggested one or two more details. definitely not perfect and might need some proof reading from a native speaker but i hope it might provide some value and a cue how to improve the paragraph